refactor(db): extract counter commands to src/db/counters.ts
Move counter types, lookup cache, CRUD operations, and yearly archive
into their own module. db.ts is now re-exports only for the counter section.
