+++ src/db.ts
@@ -1,11 +1,5 @@
 import mysql from 'mysql2/promise';
-import { createManagedLookupCache, type RefreshingLookupCache } from './db/lookupCache';
 import { getPool } from './db/pool';
-import {
-  requireTrimmedString, normalizeCommandList, isAnyCommandTakenAcrossTables,
-  runSerializedCommandWrite,
-} from './db/commandLocks';
-import type { SqlExecutor } from './db/commandLocks';
 export type { RefreshingLookupCache, ManagedLookupCacheOptions, ManagedLookupCache } from './db/lookupCache';
 export { getPool, closePool } from './db/pool';
 
@@ -137,326 +131,13 @@ export type { SqlExecutor } from './db/commandLocks';
 
 // ─── Counter commands ───────────────────────────────────────────────────────
 
-export interface DbCounter {
-  id: number;
-  trigger_command: string;
-  check_command: string;
-  message: string;
-  increment_message: string;
-  reset_yearly: boolean;
-  current_value: number;
-}
-
-export type CounterMatchType = 'trigger' | 'check';
-
-export interface DbMatchedCounter extends DbCounter {
-  matchType: CounterMatchType;
-}
-
-export class CounterNotFoundError extends Error {
-  constructor(id: number) {
-    super(`Counter not found: ${id}`);
-    this.name = 'CounterNotFoundError';
-  }
-}
-
-interface NormalizedCounterFields {
-  triggerCommand: string;
-  checkCommand: string;
-  message: string;
-  incrementMessage: string;
-}
-
-export interface UpdateCounterInput {
-  id: number;
-  triggerCommand: string;
-  checkCommand: string;
-  message: string;
-  incrementMessage: string;
-  resetYearly: boolean;
-}
-
-function normalizeCounterFields(
-  triggerCommand: string,
-  checkCommand: string,
-  message: string,
-  incrementMessage: string,
-): NormalizedCounterFields {
-  return {
-    triggerCommand: requireTrimmedString(triggerCommand, 'trigger_command').toLowerCase(),
-    checkCommand: requireTrimmedString(checkCommand, 'check_command').toLowerCase(),
-    message: requireTrimmedString(message, 'message'),
-    incrementMessage: requireTrimmedString(incrementMessage, 'increment_message'),
-  };
-}
-
-function mapCounter(row: mysql.RowDataPacket): DbCounter {
-  return {
-    id: row.id,
-    trigger_command: row.trigger_command,
-    check_command: row.check_command,
-    message: row.message,
-    increment_message: row.increment_message,
-    reset_yearly: mapBoolColumn(row.reset_yearly),
-    current_value: row.current_value,
-  };
-}
-
-interface CounterLookupCache extends RefreshingLookupCache {
-  byCommand: Map<string, DbMatchedCounter>;
-}
-
-function createEmptyCounterLookupCache(): CounterLookupCache {
-  return {
-    // Keep the fallback cache immediately stale so a new refresh can start as soon
-    // as the backoff window expires rather than waiting for the normal TTL.
-    loadedAt: 0,
-    byCommand: new Map<string, DbMatchedCounter>(),
-  };
-}
-
-const COUNTER_LOOKUP_CACHE_TTL_MS = 300_000;
-const COUNTER_LOOKUP_CACHE_REFRESH_FAILURE_BACKOFF_MS = 5_000;
-const COUNTER_LOOKUP_CACHE_REFRESH_FAILURE_MAX_BACKOFF_MS = 60_000;
-
-function buildCounterLookupCache(counters: DbCounter[]): CounterLookupCache {
-  const byCommand = new Map<string, DbMatchedCounter>();
-  const sortedCounters = [...counters].sort((left, right) => left.id - right.id);
-
-  const registerCounterCommand = (
-    normalizedCommand: string,
-    counter: DbCounter,
-    matchType: CounterMatchType,
-    commandFieldLabel: 'trigger_command' | 'check_command',
-  ): void => {
-    if (!normalizedCommand) {
-      return;
-    }
-
-    const existingCounter = byCommand.get(normalizedCommand);
-    if (existingCounter) {
-      console.warn(`[DB] Counter ${commandFieldLabel} collision: '${normalizedCommand}' is already registered (counter id=${existingCounter.id}); ignoring duplicate from counter id=${counter.id}.`);
-      return;
-    }
-
-    byCommand.set(normalizedCommand, { ...counter, matchType });
-  };
-
-  for (const counter of sortedCounters) {
-    registerCounterCommand(counter.trigger_command.trim().toLowerCase(), counter, 'trigger', 'trigger_command');
-    registerCounterCommand(counter.check_command.trim().toLowerCase(), counter, 'check', 'check_command');
-  }
-
-  return {
-    loadedAt: Date.now(),
-    byCommand,
-  };
-}
-
-const counterLookupCacheState = createManagedLookupCache<CounterLookupCache>({
-  cacheName: 'counter cache',
-  ttlMs: COUNTER_LOOKUP_CACHE_TTL_MS,
-  refreshFailureBackoffMs: COUNTER_LOOKUP_CACHE_REFRESH_FAILURE_BACKOFF_MS,
-  refreshFailureMaxBackoffMs: COUNTER_LOOKUP_CACHE_REFRESH_FAILURE_MAX_BACKOFF_MS,
-  createEmptyCache: createEmptyCounterLookupCache,
-  loadCache: async () => {
-    const counters = await getAllCounters();
-    return buildCounterLookupCache(counters);
-  },
-});
-
-async function getCounterLookupCache(): Promise<CounterLookupCache> {
-  return await counterLookupCacheState.getCache();
-}
-
-export function invalidateCounterLookupCache(): void {
-  counterLookupCacheState.invalidate();
-}
-
-export async function getAllCounters(): Promise<DbCounter[]> {
-  const [rows] = await getPool().execute<mysql.RowDataPacket[]>(
-    `SELECT id, trigger_command, check_command, message, increment_message, reset_yearly, current_value
-     FROM counter
-     ORDER BY trigger_command`,
-  );
-
-  return rows.map(mapCounter);
-}
-
-export async function findCounterByCommand(command: string): Promise<DbMatchedCounter | null> {
-  const normalizedCommand = command.trim().toLowerCase();
-  if (!normalizedCommand) {
-    return null;
-  }
-
-  const cache = await getCounterLookupCache();
-  const counter = cache.byCommand.get(normalizedCommand);
-
-  return counter
-    ? {
-      ...counter,
-    }
-    : null;
-}
-
-export async function isCounterCommandTaken(commandOrCommands: string | string[], excludeCounterId?: number): Promise<boolean> {
-  if (Array.isArray(commandOrCommands)) {
-    const normalizedCommands = normalizeCommandList(commandOrCommands);
-    if (new Set(normalizedCommands).size !== normalizedCommands.length) {
-      return true;
-    }
-  }
-
-  return await isAnyCommandTakenAcrossTables(commandOrCommands, { excludeCounterId });
-}
-
-export async function addCounter(
-  triggerCommand: string,
-  checkCommand: string,
-  message: string,
-  incrementMessage: string,
-  resetYearly: boolean,
-): Promise<void> {
-  const normalizedFields = normalizeCounterFields(triggerCommand, checkCommand, message, incrementMessage);
-  if (normalizedFields.triggerCommand === normalizedFields.checkCommand) {
-    throw new Error('Counter trigger_command and check_command must be different');
-  }
-
-  await runSerializedCommandWrite(
-    [normalizedFields.triggerCommand, normalizedFields.checkCommand],
-    undefined,
-    async (connection) => {
-      await connection.execute(
-        `INSERT INTO counter (trigger_command, check_command, message, increment_message, reset_yearly, current_value)
-         VALUES (?, ?, ?, ?, ?, 0)`,
-        [
-          normalizedFields.triggerCommand,
-          normalizedFields.checkCommand,
-          normalizedFields.message,
-          normalizedFields.incrementMessage,
-          resetYearly ? 1 : 0,
-        ],
-      );
-    },
-  );
-
-  invalidateCounterLookupCache();
-}
-
-async function counterExists(id: number, executor: SqlExecutor = getPool()): Promise<boolean> {
-  const [rows] = await executor.execute<mysql.RowDataPacket[]>(
-    'SELECT 1 FROM counter WHERE id = ? LIMIT 1',
-    [id],
-  );
-  return rows.length > 0;
-}
-
-export async function updateCounter(input: UpdateCounterInput): Promise<void> {
-  const {
-    id,
-    triggerCommand,
-    checkCommand,
-    message,
-    incrementMessage,
-    resetYearly,
-  } = input;
-
-  const normalizedFields = normalizeCounterFields(triggerCommand, checkCommand, message, incrementMessage);
-  if (normalizedFields.triggerCommand === normalizedFields.checkCommand) {
-    throw new Error('Counter trigger_command and check_command must be different');
-  }
-
-  await runSerializedCommandWrite(
-    [normalizedFields.triggerCommand, normalizedFields.checkCommand],
-    { excludeCounterId: id },
-    async (connection) => {
-      const [result] = await connection.execute<mysql.ResultSetHeader>(
-        `UPDATE counter
-         SET trigger_command = ?,
-             check_command = ?,
-             message = ?,
-             increment_message = ?,
-             reset_yearly = ?
-         WHERE id = ?`,
-        [
-          normalizedFields.triggerCommand,
-          normalizedFields.checkCommand,
-          normalizedFields.message,
-          normalizedFields.incrementMessage,
-          resetYearly ? 1 : 0,
-          id,
-        ],
-      );
-
-      if (result.affectedRows === 0 && !(await counterExists(id, connection))) {
-        throw new CounterNotFoundError(id);
-      }
-    },
-  );
-
-  invalidateCounterLookupCache();
-}
-
-export async function removeCounter(id: number): Promise<void> {
-  const [result] = await getPool().execute<mysql.ResultSetHeader>('DELETE FROM counter WHERE id = ?', [id]);
-  if (result.affectedRows === 0) {
-    throw new CounterNotFoundError(id);
-  }
-
-  invalidateCounterLookupCache();
-}
-
-export async function resetCounterCurrentValue(id: number): Promise<void> {
-  const [result] = await getPool().execute<mysql.ResultSetHeader>(
-    'UPDATE counter SET current_value = 0 WHERE id = ?',
-    [id],
-  );
-
-  if (result.affectedRows === 0 && !(await counterExists(id))) {
-    throw new CounterNotFoundError(id);
-  }
-
-  invalidateCounterLookupCache();
-}
-
-export async function incrementCounter(id: number): Promise<number> {
-  const conn = await getPool().getConnection();
-  try {
-    await conn.beginTransaction();
-    const [result] = await conn.execute<mysql.ResultSetHeader>(
-      'UPDATE counter SET current_value = current_value + 1 WHERE id = ?',
-      [id],
-    );
-    if (result.affectedRows === 0) throw new CounterNotFoundError(id);
-    const [rows] = await conn.execute<mysql.RowDataPacket[]>(
-      'SELECT current_value FROM counter WHERE id = ?',
-      [id],
-    );
-    const newValue = (rows[0] as mysql.RowDataPacket).current_value as number;
-    await conn.commit();
-    invalidateCounterLookupCache();
-    return newValue;
-  } catch (err) {
-    await conn.rollback().catch(() => {});
-    throw err;
-  } finally {
-    conn.release();
-  }
-}
-
-export async function archiveAndResetYearlyCounters(year: number): Promise<number> {
-  if (!Number.isInteger(year) || year < 2020 || year > 2100) {
-    throw new Error(`[DB] Invalid archive year: ${year}`);
-  }
-  // MySQL doesn't support parameterised column names, so the name is built from
-  // `year` which is validated to a safe integer in [2020, 2100] above.
-  const columnName = `value${year}`;
-  const [result] = await getPool().execute<mysql.ResultSetHeader>(
-    `UPDATE counter SET \`${columnName}\` = current_value, current_value = 0 WHERE reset_yearly = 1 AND \`${columnName}\` IS NULL`,
-  );
-  invalidateCounterLookupCache();
-  return result.affectedRows;
-}
+export {
+  CounterNotFoundError, invalidateCounterLookupCache,
+  getAllCounters, findCounterByCommand, isCounterCommandTaken,
+  addCounter, updateCounter, removeCounter, resetCounterCurrentValue,
+  incrementCounter, archiveAndResetYearlyCounters,
+} from './db/counters';
+export type { DbCounter, CounterMatchType, DbMatchedCounter, UpdateCounterInput } from './db/counters';
 
 // ─── Stream monitor ──────────────────────────────────────────────────────────
